+++ src/test/java/co/fineants/api/domain/portfolio/domain/entity/PortfolioTest.java
@@ -6,7 +6,6 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import org.assertj.core.api.Assertions;
-import org.assertj.core.groups.Tuple;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -14,7 +13,6 @@ import co.fineants.AbstractContainerBaseTest;
 import co.fineants.api.domain.common.count.Count;
 import co.fineants.api.domain.common.money.Expression;
 import co.fineants.api.domain.common.money.Money;
-import co.fineants.api.domain.holding.domain.dto.response.PortfolioPieChartItem;
 import co.fineants.api.domain.holding.domain.dto.response.PortfolioSectorChartItem;
 import co.fineants.api.domain.holding.domain.entity.PortfolioHolding;
 import co.fineants.api.domain.member.domain.entity.Member;
@@ -84,50 +82,6 @@ class PortfolioTest extends AbstractContainerBaseTest {
 		Money totalInvestmentAmount = Money.won(100000);
 		Expression expected = totalGainAmount.divide(totalInvestmentAmount);
 		assertThat(result).isEqualByComparingTo(expected);
-	}
-
-	@DisplayName("사용자는 포트폴리오의 파이 차트를 요청한다")
-	@Test
-	void createPieChart() {
-		// given
-		Portfolio portfolio = createPortfolio(createMember());
-		Stock stock = createSamsungStock();
-		Stock stock2 = createDongwhaPharmStock();
-		PortfolioHolding holding1 = PortfolioHolding.of(portfolio, stock, Money.won(20000L));
-		PortfolioHolding holding2 = PortfolioHolding.of(portfolio, stock2, Money.won(20000L));
-
-		LocalDateTime purchaseDate = LocalDateTime.of(2023, 9, 26, 9, 30, 0);
-		Count numShares = Count.from(5);
-		Money purchasePerShare = Money.won(10000);
-		String memo = "첫구매";
-		PurchaseHistory purchaseHistory1 = createPurchaseHistory(null, purchaseDate, numShares, purchasePerShare, memo,
-			holding1);
-
-		numShares = Count.from(5);
-		purchasePerShare = Money.won(20000);
-		PurchaseHistory purchaseHistory2 = createPurchaseHistory(null, purchaseDate, numShares, purchasePerShare, memo,
-			holding2);
-
-		holding1.addPurchaseHistory(purchaseHistory1);
-		holding2.addPurchaseHistory(purchaseHistory2);
-
-		portfolio.addHolding(holding1);
-		portfolio.addHolding(holding2);
-
-		PortfolioCalculator calculator = new PortfolioCalculator();
-		// when
-		List<PortfolioPieChartItem> items = portfolio.createPieChart(calculator);
-
-		// then
-		assertThat(items)
-			.asList()
-			.hasSize(3)
-			.extracting("name", "valuation", "totalGain")
-			.usingComparatorForType(Money::compareTo, Money.class)
-			.containsExactly(
-				Tuple.tuple("현금", Money.won(850000), Money.zero()),
-				Tuple.tuple("삼성전자보통주", Money.won(100000), Money.won(50000)),
-				Tuple.tuple("동화약품보통주", Money.won(100000), Money.zero()));
 	}
 
 	@DisplayName("사용자는 포트폴리오의 섹터 차트를 요청한다")
